pyproject.toml
[tool.poetry]
name = "bananatube"
version = "0.1.0"
description = "voice and instruments over a giga banana pipeline tube"
authors = ["BorisELG <elgarehb@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6.1,<3.10"
youtube_dl="2021.6.6"
spleeter="2.3.0"
ffmpeg-python="0.2.0"
google-cloud-storage = "^1.42.3"
pydub = "^0.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
